requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
aPRAW==0.6.9a0
async-timeout==4.0.2
attrs==22.1.0
certifi==2022.6.15
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
disnake==2.5.2
ffmpeg==1.4
Flask==2.2.2
frozenlist==1.3.1
googletrans==3.1.0a0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2022.8.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
multidict==6.0.2
praw==7.6.0
prawcore==2.3.0
psutil==5.9.1
pycparser==2.21
PyNaCl==1.5.0
python-dotenv==0.20.0
ReactivePy==2.0.0.dev0
requests==2.28.1
rfc3986==1.5.0
sniffio==1.2.0
update-checker==0.18.0
urllib3==1.26.11
websocket-client==1.3.3
Werkzeug==2.2.2
yarl==1.8.1
youtube-dl==2021.12.17
